refactor(mcp): break circular import and consolidate types
Move RegisterToolFn and ToolResult into tools/types.ts so the
dependency flows one way (registerTool.ts → types.ts, tracing.ts →
types.ts) with no cycle. Drop unused annotations field and
ToolAnnotations import.
